Fix build with gcc < 14
sqrtf was not defined in cmath

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=79700

+++ Hurrican/src/DX8Sound.cpp
@@ -145,7 +145,7 @@ void SoundManagerClass::Update3DChannel(int ch) {
     float xdiff = ((Player[0].xpos + 45) - channels[ch].xpos);
     float ydiff = ((Player[0].ypos + 45) - channels[ch].ypos);
 
-    float Abstand = sqrtf((xdiff * xdiff) + (ydiff * ydiff));
+    float Abstand = std::sqrt((xdiff * xdiff) + (ydiff * ydiff));
 
     int vol = static_cast<int>(100.0f - Abstand / 6.0f);
     if (vol < 0)
@@ -673,7 +673,7 @@ int SoundManagerClass::PlayWave3D_SDL(int x, int y, int nr) {
     float ydiff = ((Player[0].ypos + 45) - y);
 
     // DKS - converted to float:
-    float Abstand = std::sqrtf((xdiff * xdiff) + (ydiff * ydiff));
+    float Abstand = std::sqrt((xdiff * xdiff) + (ydiff * ydiff));
 
     int vol = static_cast<int>(100.0f - Abstand / 6.0f);
     if (vol < 0)
